Backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
PyMuPDF
google-generativeai
python-dotenv
